refactor: restructure OpenCode to unified mc-agent-toolkit plugin
Flatten plugins/opencode/prevent/ so that plugins/opencode/ itself is
the single OpenCode plugin. Prevent becomes a feature module under
src/prevent/. Follows Option B from the plugin architecture guide.
