+++ plugins/opencode/CHANGELOG.md
