+++ plugins/opencode/README.md
