+++ plugins/opencode/commands/mc-validate.md
