+++ plugins/opencode/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@montecarlo/opencode-prevent",
+  "name": "@montecarlo/mc-agent-toolkit",
   "version": "1.0.0",
-  "description": "Monte Carlo Prevent — detect and prevent breaking schema changes using Monte Carlo lineage and monitoring data.",
+  "description": "Monte Carlo Agent Toolkit — data observability skills and enforcement hooks for AI coding agents.",
   "type": "module",
   "main": "src/index.ts",
   "exports": {
