--- plugins/opencode/prevent/skills/prevent
@@ -1,1 +1,0 @@
-../../../../skills/prevent
