+++ plugins/opencode/skills/prevent
@@ -1,0 +1,1 @@
+../../../skills/prevent
