+++ plugins/opencode/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Monte Carlo Agent Toolkit plugin for OpenCode.
+ *
+ * Registers all feature modules. Currently includes:
+ * - MC Prevent: detect and prevent breaking schema changes
+ */
+export { McPrevent } from "./prevent/index";
+export { default } from "./prevent/index";
