+++ plugins/opencode/src/prevent/cache.ts
