+++ plugins/opencode/src/prevent/detect.ts
