+++ plugins/opencode/src/prevent/index.ts
