+++ plugins/opencode/tests/prevent/helpers.ts
