+++ plugins/opencode/tsconfig.json
